% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamline.r
\name{clean_aggregation}
\alias{clean_aggregation}
\title{clean_aggregation}
\usage{
clean_aggregation(csv_file, country)
}
\arguments{
\item{csv_file}{input csv_file}

\item{country}{country for which the report is issued}
}
\value{
None
}
\description{
general clean function for J.H website gather daily data
for data already aggregated/filtered
}
\examples{
# Clean the dataset: "03-04-2021.csv"
clean_aggregation('03-04-2021.csv')
}
