% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamline.r
\name{confirmed_death_summarize_state}
\alias{confirmed_death_summarize_state}
\title{Create a report with confirmed cases and deaths
for each State}
\usage{
confirmed_death_summarize_state(dataframe)
}
\arguments{
\item{dataframe}{input (cleaned) dataframe - contains Admin2}
}
\value{
summarized_dataframe
}
\description{
create the daily report
}
\examples{
# create a report: confirmed_death_summarize_county(dataframe)
 
}
