% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamline.r
\name{create_full_link}
\alias{create_full_link}
\title{Create Link function}
\usage{
create_full_link(date = "01-01-2021", aggregated = TRUE)
}
\arguments{
\item{date.}{(date before which the data is retrieved)}

\item{aggregated.}{(data is available as already aggregated - US ONLY or raw - all countries )}
}
\value{
An url string - complete link
}
\description{
Concatenate date and repository link
}
\examples{
# Create the url based on date : 03-04-2021
create_link('03-04-2021')
}
