% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamline.r
\name{create_out_name_based_on_date}
\alias{create_out_name_based_on_date}
\title{Create filename based on date}
\usage{
create_out_name_based_on_date(date, aggregated = TRUE)
}
\arguments{
\item{date}{}
}
\value{
None
}
\description{
create the out_folder format
}
\examples{
# save a file: create_out_name_based_on_date("03-04-2021")
 
}
