% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamline.r
\name{create_out_name_based_on_url}
\alias{create_out_name_based_on_url}
\title{Create filename based on url}
\usage{
create_out_name_based_on_url(url, aggregated = FALSE)
}
\arguments{
\item{url}{dataset url (from which is extracted the output filename)}

\item{aggregated}{boolean flag to indicate whether data was already aggregated}
}
\value{
dest_file_name
}
\description{
create the out_folder format
}
\examples{
# create output filename: create_out_name_based_on_url("https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_daily_reports_us/03-04-2021.csv")
 
}
