% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamline.r
\name{download_link}
\alias{download_link}
\title{Download Link function}
\usage{
download_link(url, aggregated = TRUE)
}
\arguments{
\item{url}{}
}
\value{
None
}
\description{
Download link available on J.H website
}
\examples{
# Download the url : "https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_daily_reports_us/03-04-2021.csv"
download_link('03-04-2021')
}
