% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamline.r
\name{JH_confirmed_death_summarize}
\alias{JH_confirmed_death_summarize}
\title{Create a report with confirmed cases and deaths
averaged over counties for each State}
\usage{
JH_confirmed_death_summarize(dataframe, aggregated = FALSE)
}
\arguments{
\item{dataframe}{input (cleaned) dataframe}

\item{aggregated}{boolean flag to indicate whether data was already aggregated}
}
\value{
summarized_dataframe
}
\description{
create the daily report
}
\examples{
# create a report: confirmed_death_summarize_county(dataframe)
 
}
