% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamline.r
\name{util_confirmed_death_summarize_state}
\alias{util_confirmed_death_summarize_state}
\title{Create a report with confirmed cases and deaths
for each State (Helper function)}
\usage{
util_confirmed_death_summarize_state(dataframe, date = "01-01-2021")
}
\arguments{
\item{dataframe}{input (cleaned) dataframe - contains Admin2}
}
\value{
summarized_dataframe
}
\description{
create the daily report
}
\examples{
# create a report: confirmed_death_summarize_county(dataframe)
 
}
