% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamline.r
\name{util_JH_general_clean}
\alias{util_JH_general_clean}
\title{general cleaner for JH}
\usage{
util_JH_general_clean(dataframe, country)
}
\arguments{
\item{dataframe}{queried dataframe}

\item{country}{filtered on}
}
\value{
dataframe
}
\description{
create a general clean
}
\examples{
# create output filename: create_out_name_based_on_date("https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_daily_reports_us/03-04-2021.csv")
 
}
