% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamline.r
\name{clean_us_data2}
\alias{clean_us_data2}
\title{Clean_us_data}
\usage{
clean_us_data2(csv_file, aggregated = TRUE)
}
\arguments{
\item{csv_file}{}
}
\value{
None
}
\description{
clean J.H website to limit correct data in the US
}
\examples{
# Filter the dataset to US entries: "03-04-2021.csv"
clean_us_data('03-04-2021.csv')
}
