% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamline.r
\name{clean_us_data2}
\alias{clean_us_data2}
\title{Clean_us_data}
\usage{
clean_us_data2(date = "01-01-2021", country = "US", aggregated = TRUE)
}
\arguments{
\item{date}{queried date for analysis}

\item{country}{queried country for analysis}

\item{aggregated}{boolean flag to indicate whether data was already aggregated}
}
\value{
None
}
\description{
clean J.H website to limit data in the US
}
\examples{
# Filter the dataset to US entries: "03-04-2021.csv"
clean_us_data2('03-04-2021.csv')
}
