% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamline.r
\name{create_full_link}
\alias{create_full_link}
\title{Create Link function}
\usage{
create_full_link(date = "01-01-2021", aggregated = TRUE)
}
\arguments{
\item{date.}{queried date for analysis}

\item{aggregated.}{boolean flag to indicate whether data was already aggregated}
}
\value{
An url string - complete link
}
\description{
Concatenate date and repository link
}
\examples{
# Create the url based on date : 01-01-2021
create_link('03-04-2021')
}
