% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamline.r
\name{create_out_name_based_on_date}
\alias{create_out_name_based_on_date}
\title{Create filename based on date}
\usage{
create_out_name_based_on_date(date = "01-01-2021", aggregated = TRUE)
}
\arguments{
\item{date}{queried date for analysis}

\item{aggregated}{boolean flag to indicate whether data was already aggregated}
}
\value{
dest_file_name
}
\description{
create the out_folder format
}
\examples{
# create output filename: create_out_name_based_on_date("https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_daily_reports_us/03-04-2021.csv")
 
}
