% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamline.r
\name{create_out_name_based_on_url}
\alias{create_out_name_based_on_url}
\title{Create filename based on url}
\usage{
create_out_name_based_on_url(url, aggregated = TRUE)
}
\arguments{
\item{date}{}
}
\value{
None
}
\description{
create the out_folder format
}
\examples{
# save a file: create_out_name_based_on_url("https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_daily_reports_us/03-04-2021.csv")
 
}
